studybud/requirement.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-mssql-backend==2.8.1
django-pyodbc-azure==2.1.0.0
djangorestframework==3.14.0
mssql-django==1.3
Pillow==10.0.0
pyodbc==4.0.39
pytz==2023.3
sqlparse==0.4.2
typing_extensions==4.7.1
tzdata==2022.1
